requirements.txt
# Requirements for YouTube downloader and vocal remover project

# YouTube downloader dependencies
pytube>=12.1.0
ffmpeg-python>=0.2.0
yt-dlp>=2023.3.4

# Vocal remover dependencies (using Demucs instead of Spleeter)
demucs>=4.0.0
torchaudio>=0.9.0
torch>=1.9.0

# Additional useful audio processing libraries
librosa>=0.9.1
pydub>=0.25.1
matplotlib>=3.5.0
moviepy==1.0.3
numpy>=1.20.0

# Environment variable management
python-dotenv>=0.21.0

# API requests for Eleven Labs
requests>=2.28.0